refactor(android): Placesチャネル処理を分離
